// mqtt-handler/src/type.ts
export type JSONDocv1= {
    version: string;
    rpm: number;
    temp1: number;
    temp2: number;
    temp3: number;
    temp4: number;
    lube_oil_pressure: number;
    water_presence: boolean;
    water_flow_volume_rate_pulse: number;
    fuel_sensor_level: number;
    engine_switch: number;
    engine_status: number;
}